Make CodecFactory class more cross-platform
Add a static platform detection class that detects Windows, Mac and Linux.

Only register Windows specific codecs if running on Windows.

Provide a RegisterDefaultCodec method in the Codec Factory that allows customization of the default codec. This makes use of the DefaultCodecAction delegate.

These changes mean that on startup for e.g. Mac, only cross-platform codecs are registered in the Codec Factory. It is up to the user of CSCore to call OSXAudio.RegisterCodecs() which then registers the OSX CoreAudio api with the Codec Factory.