PolicyConfig class to manipulate Windows audio devices
This class enables the programmer to manipulate Windows settings for audio devices, such as the default device for each role. It's rather messy due to the fact that the COM Interface IID changes with a considerable frequency, hence we need to try to support every possibility, falling back to an unknown interface (https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-dcom/2b4db106-fb79-4a67-b45f-63654f19c54c) as a last resort. This class is basically a mashup of many different implementations of the same feature across public repositories, extracting, to my best judgment, the best of each. It's most probably not following the project code standards so I'm kinda expecting for someone to help me on that :)

Sorry for any inconvenience!