Builds with .NET8.  Basic audio player test works. has not been fully tested.
The codebase uses a hacky post-processing step inherited for the framework 4.5 days that I have hacked further. It *seems* to be working but is somewhat fragile and thre are probably better ways to do it in modern .NET

This should be considered an experimental build. If the old Framework 4.5 is adaqaute for your needs I recommend staying with that.